return methods added, rm unused pkgs, getWh method updated

+++ ubs/src/main/java/com/ubs/ubs/repository/CancelRepository.java
@@ -1,12 +1,11 @@
 package com.ubs.ubs.repository;
 
 import org.springframework.data.jpa.repository.JpaRepository;
-import org.springframework.data.repository.CrudRepository;
 import org.springframework.stereotype.Repository;
 
 import com.ubs.ubs.model.Cancel;
 
 @Repository
-public interface CancelRepository extends JpaRepository<Cancel,Long> {
+public interface CancelRepository extends JpaRepository<Cancel,Integer> {
 
 }
