+++ ubs/src/main/java/com/ubs/ubs/repository/GoodinRepository.java
@@ -1,13 +1,12 @@
 package com.ubs.ubs.repository;
 
 import org.springframework.data.jpa.repository.JpaRepository;
-import org.springframework.data.repository.CrudRepository;
 import org.springframework.stereotype.Repository;
 
 import com.ubs.ubs.model.Goodin;
 
 @Repository
-public interface GoodinRepository extends JpaRepository<Goodin,Long>{
+public interface GoodinRepository extends JpaRepository<Goodin,Integer>{
 	Goodin findById(int i);
 
 }
